fix: prevent exception when get_foreground_process_name called in multiplexer
`pane:get_foreground_process_name()` is documented as returning `nil` when run in the multiplexer of via SSH.
This means that the `foreground_process` variable may be `nil` when the conditional that is searching it as if it were a string will instead raise an exception as it cannot call that operation on `nil`.

To resolve the error, this patch checks for a non-nil value at the start of the conditional.